Add files via upload
Primary upload of the UI and SERVER R files for the project. Modifications to ISS Critical Levels and Financial Deficits pages -- working on addition of navbarPanel for 'Individual' and 'Compare' on Financial Deficits, checkbox/graph for 'Compare' (checkbox exists but not reactive yet, not sure how to create multiple-line graph for all checked service options).