Add files via upload
Got individual checkbox options to work -- if we try to select 0 or > 1 checkbox, we receive an error in the console (Error in switch: EXPR must be a length 1 vector). Using Tom's arlington server file to try to use subset function, but may need updatecheckboxgroupInput and update session information instead. Trying to come up with a way to save col_names ("ISS1"..."ISS71") and actual service names to separate lists that can be co-referenced to save space in UI code (note*: c( ) concatenates, does not retain individual information).