Add files via upload
Working on the financials to get risk ratios for deficit/requirement. Finally got everything running smoothly - need to create lists of labels (under-, over-, met), separate the ratiorows dictionary into tuples to be downloaded as a .csv file, need to invert all percentages to be opposite of what they currently are (negatives > positives, and vice versa).