Add files via upload
Working to "add in" the missing ISSs. Still missing 115, 116, 803, 308, 903-906. Got deficits for the found ISSs (414, 304, 309, 103, 104) and need to work on ratios tomorrow. Used deficit_match to write a .csv file with all the deficit/requirement ratios and updated issforcritlevels with all this new information about the financial data for the found ISSs. Tomorrow, will create bubble chart for risk (after all ratios are good to go) and draw up design for website. Meeting with Claire from 11:30-3pm for coding sesh.