Add files via upload
Here are all the files for the final version of the application. There's still quite a bit more that I could do that would make the program more efficient -- condensing all the files into 3 or 4 .csv files (namely triplefinancials.csv, finishedcomments2.csv, compiledH.csv) would be a start, because a lot of the data could be handled in R and used to create new dataframes within the program locally without importing extra csv files. With the timeline being 4-5 weeks, once I realized that I needed to stop importing csv files, I didn't have time to go back and re-code everything. I did what I could to condense onto triplefinancials.csv, but it's obvious that riskfeed, riskflip, and isscritlevels2 are not necessary on their own. Hindsight is 20/20, but with the work being so fast-paced and limited, I think this was great for a first attempt at an R Shiny application and I know what I will improve on for the next time.